+++ yggdrasil_decision_forests/metric/ranking_ap.cc
@@ -13,29 +13,37 @@
  * limitations under the License.
  */
 
-#include "yggdrasil_decision_forests/metric/ranking_mrr.h"
+#include "yggdrasil_decision_forests/metric/ranking_ap.h"
 
 #include <algorithm>
 #include <vector>
 
 #include "yggdrasil_decision_forests/metric/ranking_utils.h"
+
 namespace yggdrasil_decision_forests {
 namespace metric {
 
-MRRCalculator::MRRCalculator(const int truncation) : truncation_(truncation) {}
+APCalculator::APCalculator(const int truncation) : truncation_(truncation) {}
 
-double MRRCalculator::MRR(
+double APCalculator::AP(
     const std::vector<RankingLabelAndPrediction>& group) const {
   auto mutable_group = group;
   std::sort(mutable_group.begin(), mutable_group.end(),
             OrderDecreasingPrediction);
   const int max_rank = std::min(truncation_, static_cast<int>(group.size()));
+  double average_precision = 0.0;
+  double num_relevant = 0;
   for (int rank = 0; rank < max_rank; rank++) {
     if (mutable_group[rank].label > 0.5) {
-      return 1.0 / (rank + 1.0);
+      num_relevant++;
+      average_precision += num_relevant / (rank + 1);
     }
   }
-  return 0.0;
+  if (num_relevant > 0) {
+    return average_precision / num_relevant;
+  } else {
+    return 0.0;
+  }
 }
 
 }  // namespace metric